separate test file for code_utils functions
